renamed training script
